% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualSeasonalMeans.R
\name{AnnualSeasonalMeans}
\alias{AnnualSeasonalMeans}
\title{Calculate the Annual and Seasonal Summary Values from Monthly ClimateNA Variables}
\usage{
AnnualSeasonalMeans(
  item,
  var = c("Tmax", "Tmin", "PPT", "DD5", "NFFD", "PAS", "CMI")
)
}
\arguments{
\item{item}{CSV file containing monthly (NOT primary monthly) ClimateNA variables.}

\item{var}{Climate variables desired for mean/sum calculations based on annual and seasonal distributions (seasons based on months, not true seasons).  Options include : Tmax, Tmin, Tave, PPT, Rad, DD_5, DD5, DD_18, DD18, NFFD, PAS, Eref, CMD, RH, CMI.  Mean values are calculated for temperature, solar radiation, and relative humidity.  Summed values are calculated for all other variables. Default value indicates only Tmax, Tmin, PPT, DD5, NFFD, PAS, and CMI.}
}
\value{
An expanded data frame
}
\description{
Calculate the Annual and Seasonal Summary Values from Monthly ClimateNA Variables
}
\examples{
### Processing a single file
#files <- list.files(pattern='*csv$')
#site <- read.csv(files[1])
#annSeaMean <- AnnualSeasonalMeans(site,c('Tmax','Tmin','PPT','CMI'))

## Processing multiple files
files <- list.files(pattern='*csv$')
for(i in files){
  item <- read.csv(i)
  item <- AnnualSeasonalMeans(item)
  write.csv(item,i)
}
}
