% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateSummary.R
\name{climateSummary}
\alias{climateSummary}
\title{Summarize climate variables by Year}
\usage{
climateSummary(file, startYear, endYear, outdir = dirname(file))
}
\arguments{
\item{file}{The climate output from ClimateNA as CSV file format}

\item{startYear}{numeric. The first year of the time frame}

\item{endYear}{numeric. The last year of the time frame}

\item{outdir}{character. The output directory for the summary file}
}
\value{
A CSV file consisting of the summary data
}
\description{
Climate data is summarized based on the variable's mean and standard deviation values for each year and is outputted to a summarized table.
}
\examples{
## Input a data file consisting of monthly climate data projections from 2000 to 2050
#file <- read.csv('data.csv')
#climateSummary(file,2000,2050)
}
\author{
Michael Burnett - UBC Faculty of Forestry
}
