% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVtoTIFF.R
\name{CSVtoTIFF}
\alias{CSVtoTIFF}
\title{Rasterize Climate data from ClimateNA}
\usage{
CSVtoTIFF(
  Longitude,
  Latitude,
  Value,
  filename,
  crsref = "+proj=longlat +datum=WGS84",
  outdir = getwd()
)
}
\arguments{
\item{Longitude}{The data frame column including the Longitudinal data}

\item{Latitude}{The data frame column including the Latitudinal data}

\item{Value}{numeric. The data frame column including the raster pixel values}

\item{filename}{character. Output file name. Must include file type (example ".tif")}

\item{crsref}{character. Description of the Coordinate Reference System (map projection) in PROJ.4, WKT or authority:code notation}

\item{outdir}{character. Output directory for the raster file}
}
\value{
A raster representative of the climate data acquired through ClimateNA
}
\description{
Rasterize Climate data from ClimateNA
}
\examples{
#data <- read.csv('climateNA_monthlyDataOutput.csv')
#CSVtoTIFF(data$Longitude,data$Latitude,data$Tmin,'annual_Tmin.tif')

}
\author{
Michael Burnett - UBC Faculty of Forestry
}
