% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist30YClimateNA.R
\name{hist30YClimateNA}
\alias{hist30YClimateNA}
\title{Access 30 Years Normal Historical ClimateNA Data}
\usage{
hist30YClimateNA(file, dateR, tFrame, exe, direc = getwd())
}
\arguments{
\item{file}{the prepared DEM as a CSV file. Full file name is required.}

\item{dateR}{the historical date range for the climate data. Make sure it is in the 30 years format of 'XXX1_XXX0'. Underscore is necessary, as opposed to dash.}

\item{tFrame}{the averaged time frame of each climate variable. Use 'M' for monthly, 'A' for annual, and 'S' for seasonal}

\item{exe}{Full name of the ClimateNA exe file. Make sure that your working directory is the folder in which this file is located in.}

\item{direc}{a character vector of full path name to the directory where the ClimateNA app is located.}
}
\value{
A CSV file with historical climate data in the current directory
}
\description{
Access 30 Years Normal Historical ClimateNA Data
}
\examples{
#setwd("E:/Climatena_v721");getwd() # Set up location for the application
#exe <- "E:/Climatena_v721/ClimateNA_v7.21.exe"
#files <- list.files(pattern = '*.csv$')
#hist30YClimateNA(files[1],'1991_2020','M',exe)
}
