% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demToCSV.R
\name{demToCSV}
\alias{demToCSV}
\title{Convert a Raster DEM file to a Useable CSV that can be inputted into ClimateNA}
\usage{
demToCSV(file, outdir = getwd())
}
\arguments{
\item{file}{A TIFF file at the desired spatial resolution.}

\item{outdir}{The output directory for the CSV file.}
}
\value{
A CSV file representing the inputted DEM
}
\description{
Convert a Raster DEM file to a Useable CSV that can be inputted into ClimateNA
}
\examples{
#files <- list.files(pattern='*.tif$')
#roiDEM <- files[3]
#demToCSV(roiDEM)
}
