% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmSepCols.R
\name{gcmSepCols}
\alias{gcmSepCols}
\title{Separate Components of ClimateNA File Names}
\usage{
gcmSepCols(files)
}
\arguments{
\item{files}{A list of ClimateNA future projection outputs.}
}
\value{
List
}
\description{
Separate Components of ClimateNA File Names
}
\examples{
files <- list.files()
models <- gcmSepCols(files)
}
