% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasToPts.R
\name{rasToPts}
\alias{rasToPts}
\title{List of Rasters To Shapefile Points}
\usage{
rasToPts(files)
}
\arguments{
\item{files}{A list of raster files.}
}
\value{
Point Featureclass Shapefile
}
\description{
Use a list of rasters in a directory (ASCII output files from ClimateNA) to produce point feature class shapefiles.
}
\examples{
files <- list.files(pattern='*asc$') # Access all ASCII raster files in directory
rasToPts(files) # Produce shapefile outputs
}
\author{
Michael Burnett - UBC Faculty of Forestry
}
